- 新增API每日调用统计表和增量聚合检查点   - 使用log_id分批聚合原始日志，每批最多处理5000条   - 每日统计、累计调用量和检查点在同一事务内更新   - 增加多实例聚合行锁，避免重复统计   - 新增GET /api/v1/stat/dashboard统一统计接口   - Dashboard改为读取聚合表，不再扫描原始日志   - 增加Caffeine本地缓存，统计结果缓存30秒   - 前端统计页面由多次并发请求改为单次Dashboard请求   - 将前端统计类型统一移动到types/module/stat.ts   - 统计数量类型由int升级为long   - API或用户删除时同步清理相关聚合数据   - 新增增量聚合和Dashboard服务测试

+++ server/src/main/java/ee/zxz/helloapi/domain/DTO/ApiTodayArray.java
@@ -3,7 +3,7 @@ package ee.zxz.helloapi.domain.DTO;
 
 public class ApiTodayArray {
     private String name;
-    private Integer count;
+    private Long count;
 
     public String getName() {
         return name;
@@ -13,11 +13,11 @@ public class ApiTodayArray {
         this.name = name;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 
-    public void setCount(Integer count) {
+    public void setCount(Long count) {
         this.count = count;
     }
 }
